Refactor: move product routes to separate file
