+++ server/src/main.rs
@@ -1,16 +1,13 @@
 use axum::body::{boxed, Body};
-use axum::extract::{self, Extension};
 use axum::http::{Response, StatusCode};
 use axum::{
     response::IntoResponse,
     routing::{get, post},
-    Json, Router,
+    Router,
 };
 use clap::Parser;
-use entity::{product, sea_orm};
 use migration::{Migrator, MigratorTrait};
-use product::Entity as Product;
-use sea_orm::{prelude::*, Database, JsonValue, QueryOrder, Set};
+use sea_orm::Database;
 use std::env;
 use std::net::{IpAddr, Ipv6Addr, SocketAddr};
 use std::path::PathBuf;
@@ -22,9 +19,8 @@ use tower_http::services::ServeDir;
 use tower_http::trace::TraceLayer;
 
 mod dtos;
-use dtos::ProductDto;
 mod errors;
-use errors::AppError;
+mod product_routes;
 
 // Setup the command line interface with clap.
 #[derive(Parser, Debug)]
@@ -68,8 +64,8 @@ async fn main() {
 
     let api_routes = Router::new()
         .route("/hello", get(hello))
-        .route("/products", get(list_products))
-        .route("/product", post(insert_product));
+        .route("/products", get(product_routes::list))
+        .route("/product", post(product_routes::insert));
 
     let app = Router::new()
         .nest("/api", api_routes)
@@ -122,60 +118,4 @@ async fn main() {
 
 async fn hello() -> impl IntoResponse {
     "hello from server!"
-}
-
-async fn list_products(Extension(ref conn): Extension<DatabaseConnection>) -> Json<Vec<JsonValue>> {
-    Json(
-        Product::find()
-            .filter(product::Column::Stock.gt(0))
-            .order_by_desc(product::Column::Stock)
-            .into_json()
-            .all(conn)
-            .await
-            .unwrap(),
-    )
-}
-
-async fn insert_product(
-    extract::Json(product_dto): extract::Json<ProductDto>,
-    Extension(ref conn): Extension<DatabaseConnection>,
-) -> Result<Json<ProductDto>, AppError> {
-    // validate stock
-    let stock = product_dto.stock;
-    if stock == 0 {
-        return Err(AppError::BadInput("stock must be greater than 0"));
-    }
-    // validate price
-    let price = product_dto.price;
-    if price == 0 {
-        return Err(AppError::BadInput("price must be greater than 0"));
-    }
-    // validate name
-    let name = product_dto.name.trim();
-    if name.len() == 0 {
-        return Err(AppError::BadInput("name can't be empty"));
-    }
-    // validate description
-    let description = product_dto.description.and_then(|s| {
-        let trimmed = s.trim();
-        if trimmed.len() > 0 {
-            Some(trimmed.to_string())
-        } else {
-            None
-        }
-    });
-    // TODO: get seller id from cookies
-    let product = product::ActiveModel {
-        stock: Set(stock),
-        price: Set(price),
-        name: Set(name.to_string()),
-        description: Set(description),
-        ..Default::default()
-    };
-
-    let product = product.insert(conn).await.expect("could not insert post"); // TODO
-
-    let new_product_dto = ProductDto::from_entity(product, conn).await;
-
-    Ok(Json(new_product_dto))
 }
